+++ src/index.test.js
@@ -1,5 +1,5 @@
-import Form from './index.js'
 import { reaction, isObservable, toJS } from 'mobx'
+import Form from './index'
 import _ from 'lodash/fp'
 
 require('util').inspect.defaultOptions.depth = null
@@ -7,40 +7,45 @@ require('util').inspect.defaultOptions.depth = null
 describe('Form', () => {
   let dispose = _.noop
   let form = null
-  let fields = {
-    location: {
+
+  beforeEach(() => {
+    dispose()
+    form = Form({
       fields: {
-        'country.state': {
-          fields: {
-            zip: {
-              validator: () => 'Invalid zip',
-            },
+        location: {
+          value: {
+            'country.state': { zip: '07016' },
+            addresses: [{ street: 'Meridian', tenants: ['John'] }],
           },
-        },
-        addresses: {
-          items: {
-            street: {},
-            tenants: {
+          fields: {
+            'country.state': {
+              label: 'Dotted field name',
+              fields: {
+                zip: {
+                  validator: () => 'Invalid zip',
+                },
+              },
+            },
+            addresses: {
+              label: 'Array field',
               items: {
-                name: { validator: () => 'Invalid name' },
-                age: {},
+                label: 'Item field is a record',
+                fields: {
+                  street: {},
+                  tenants: {
+                    label: 'Array field',
+                    items: {
+                      label: 'Item field is a primitive',
+                      validator: () => 'Invalid tenant',
+                    },
+                  },
+                },
               },
             },
           },
         },
       },
-      value: {
-        'country.state': { zip: '07016' },
-        addresses: [
-          { street: 'Meridian', tenants: [{ name: 'John', age: 53 }] },
-        ],
-      },
-    },
-  }
-
-  beforeEach(() => {
-    dispose()
-    form = Form({ fields })
+    })
   })
 
   it('Initialize nested fields with correct values', () => {
@@ -48,7 +53,7 @@ describe('Form', () => {
     let street = form.getField(path)
     expect(street).not.toBeUndefined()
     expect(street.value).toBe(_.get(path, form.value))
-    path = 'location.addresses.0.tenants.0.name'
+    path = 'location.addresses.0.tenants.0'
     let name = form.getField(path)
     expect(name).not.toBeUndefined()
     expect(name.value).toBe(_.get(path, form.value))
@@ -59,13 +64,13 @@ describe('Form', () => {
   })
 
   it('Two-way binding between nested field value and parent value', () => {
-    let path = 'location.addresses.0.tenants.0'
-    let name = form.getField(`${path}.name`)
+    let path = 'location.addresses.0.tenants'
+    let name = form.getField(`${path}.0`)
     let parent = _.get(path, form.value)
     name.value = 'Alex'
-    expect(parent.name).toBe(name.value)
+    expect(parent[0]).toBe(name.value)
     parent.name = 'Meridian'
-    expect(name.value).toBe(parent.name)
+    expect(name.value).toBe(parent[0])
   })
 
   it('Add new array field', () => {
@@ -82,26 +87,25 @@ describe('Form', () => {
     expect(_.last(_.get(path, form.value))).toBe(added)
 
     // Field was added correctly
-    let name = form.getField(`${path}.1.street`)
-    expect(name).not.toBeUndefined()
-    expect(name.value).toBe('Washington')
+    let street = form.getField('location.addresses.1.street')
+    expect(street).not.toBeUndefined()
+    expect(street.value).toBe('Washington')
   })
 
   it('Add new nested array field', () => {
     let path = 'location.addresses.0.tenants'
     let tenants = form.getField(path)
-    let value = { name: 'Paul' }
+    let value = 'Paul'
     tenants.add(value)
     let added = _.last(tenants.value)
 
     // Value was added correctly
     expect(added).not.toBeUndefined()
-    expect(isObservable(added)).toBe(true)
-    expect(toJS(added)).toEqual(value)
+    expect(added).toEqual(value)
     expect(_.last(_.get(path, form.value))).toBe(added)
 
     // Field was added correctly
-    let name = form.getField(`${path}.1.name`)
+    let name = form.getField(`${path}.1`)
     expect(name).not.toBeUndefined()
     expect(name.value).toBe('Paul')
   })
@@ -129,7 +133,9 @@ describe('Form', () => {
     let location = form.getField('location')
 
     // Remove object field
-    let path = 'addresses.0.tenants.0.name'
+    let path = 'addresses.0.tenants.0'
+    expect(location.getField(path)).not.toBeUndefined()
+    expect(_.get(path, location.value)).not.toBeUndefined()
     location.remove(path)
     expect(location.getField(path)).toBeUndefined()
     expect(_.get(path, location.value)).toBeUndefined()
@@ -148,7 +154,7 @@ describe('Form', () => {
     let name = null
 
     beforeEach(() => {
-      name = form.getField('location.addresses.0.tenants.0.name')
+      name = form.getField('location.addresses.0.tenants.0')
     })
 
     it('errors | isValid', () => {
@@ -166,7 +172,7 @@ describe('Form', () => {
         )
       )
       name.validate()
-      expect(errors).toHaveBeenCalledWith('Invalid name')
+      expect(errors).toHaveBeenCalledWith('Invalid tenant')
       expect(isValid).toHaveBeenCalledWith(false)
       form.reset()
       expect(errors).toHaveBeenCalledWith([])
@@ -194,7 +200,7 @@ describe('Form', () => {
       expect(name.value).toEqual('John')
 
       // Resets subfields
-      let state = form.getField('location."country.state"')
+      let state = form.getField('location.["country.state"]')
       state.add({ newField: { label: 'New name' } })
       expect(state.getField('newField')).not.toBeUndefined()
       state.reset()
@@ -209,13 +215,13 @@ describe('Form', () => {
       expect(addresses.fields.length).toBe(1)
       expect(addresses.getField('0.street').value).toBe('Meridian')
       expect(addresses.value).toEqual([
-        { street: 'Meridian', tenants: [{ name: 'John', age: 53 }] },
+        { street: 'Meridian', tenants: ['John'] },
       ])
     })
 
     it('validate()', () => {
       expect(name.validate()).toEqual({
-        'location.addresses.0.tenants.0.name': 'Invalid name',
+        'location.addresses.0.tenants.0': 'Invalid tenant',
       })
     })
 
@@ -230,7 +236,7 @@ describe('Form', () => {
       name.clean()
 
       // Cleans subfields
-      let state = form.getField('location."country.state"')
+      let state = form.getField('location.["country.state"]')
       state.add({ newField: { label: 'New name' } })
       expect(state.getField('newField')).not.toBeUndefined()
       state.clean()
@@ -244,7 +250,7 @@ describe('Form', () => {
       expect(addresses.isDirty).toEqual(false)
       expect(addresses.getField('1.street')).not.toBeUndefined()
       expect(addresses.value).toEqual([
-        { street: 'Meridian', tenants: [{ name: 'John', age: 53 }] },
+        { street: 'Meridian', tenants: ['John'] },
         { street: 'Laurence' },
       ])
     })
@@ -255,8 +261,7 @@ describe('Form', () => {
       expect(form.getSnapshot()).toEqual({
         'location.country.state.zip': '07016',
         'location.addresses.0.street': 'Meridian',
-        'location.addresses.0.tenants.0.name': 'John',
-        'location.addresses.0.tenants.0.age': 53,
+        'location.addresses.0.tenants.0': 'John',
       })
     })
 
@@ -269,7 +274,7 @@ describe('Form', () => {
           addresses: [
             {
               street: 'Meridian',
-              tenants: [{ name: 'John', age: 53 }],
+              tenants: ['John'],
             },
           ],
         },
@@ -302,7 +307,7 @@ describe('Form', () => {
       form.reset()
       expect(form.getPatch()).toEqual({})
 
-      let state = form.getField('location."country.state"')
+      let state = form.getField('location.["country.state"]')
       state.add({ capital: {} })
       expect(form.getPatch()).toEqual({})
       state.getField('capital').value = 'Tallahase'
@@ -348,7 +353,7 @@ describe('Form', () => {
       expect(form.isValid).toBe(true)
       expect(form.validate()).toEqual({
         'location.country.state.zip': 'Invalid zip',
-        'location.addresses.0.tenants.0.name': 'Invalid name',
+        'location.addresses.0.tenants.0': 'Invalid tenant',
       })
       expect(form.isValid).toBe(false)
     })