+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.3.0
-* Add nested array fields support
+* Add nested fields support
+* Add array fields support
 
 # 0.2.0
 * Export validators